迁移pdp页compare模块、ifa所有组件 先放会gemebuild目录中
