+++ apps/gemebuild/package.json
@@ -28,6 +28,8 @@
     "d3-scale": "^4.0.2",
     "lodash.sortby": "^4.7.0",
     "mixpanel-browser": "^2.47.0",
+    "moment": "^2.30.1",
+    "react-countdown": "^2.3.5",
     "react-scroll": "^1.9.0",
     "react-simple-maps": "^3.0.0",
     "swiper": "^10.2.0"
