+++ apps/gemebuild/src/i18n-pages/ifa/components/IFAHeroSection.tsx
@@ -1,13 +1,12 @@
 import Image from 'next/image'
 import CountDown from './CountDown'
 import clsx from 'clsx'
-import { ILinkComponent } from '../../../../apps/gemebuild/src/contexts/link-context'
 import { PropsWithChildren, ReactNode } from 'react'
-import { LocaleType } from '../reviews'
 import Video from './Video'
-import { Divider } from '../divider'
 import { Button } from './Button'
-import ImageZoomViewer from '../client-components/ImageZoomViewer'
+import { Divider, LocaleType } from 'ui'
+import { ILinkComponent } from '../../../contexts/link-context'
+import ImageZoomViewer from 'ui/src/client-components/ImageZoomViewer'
 
 export interface IIFAHeroSectionProps {
   noticeTitle: ReactNode
@@ -154,7 +153,8 @@ export function IFAHeroSection({
                   videoConfig={{
                     autoPlay: true,
                     src: 'https://www-geme-bio-us.s3.us-west-1.amazonaws.com/user-videos/ifa-page-exampl-video.mp4',
-                    poster: '/assets/images/ifa/poster-images/ifa-example-video-poster.png',
+                    poster:
+                      '/assets/images/ifa/poster-images/ifa-example-video-poster.png',
                     className:
                       'w-40 h-40 md:w-56 md:h-56 lg:w-[259px] lg:h-64 xl:w-56 xl:h-56 2xl:w-full 2xl:h-[292px] max-h-[292px] ',
                   }}
