+++ apps/gemebuild/src/i18n-pages/ifa/components/index.ts
@@ -3,3 +3,4 @@ export * from './WinnersNamesTicker'
 export * from './VideoList'
 export * from './LocationSection'
 export * from './ComposterSurvey'
+export * from './subscribe-geme-terra-2'
