--- packages/ui/src/compare-section/index.ts
@@ -1,1 +1,0 @@
-export * from './CompareSection'
