+++ packages/ui/src/index.tsx
@@ -27,11 +27,8 @@ export * from './Inline-player'
 export * from './ui-components'
 export * from './product-components'
 export * from './how-it-works'
-export * from './compare-section'
 export * from './waterfall-flow'
-export * from './ifa'
 export * from './divider'
-export * from './subscribe-geme-terra-2'
 export * from './user-stories-videos'
 export * from './loading-components'
 export * from './hero-section-video'
